Revert "Redirect requests to root servlet without trailing slash"
This reverts commit 13e931b5fec37e19f6fb97ed0594be4f9437184f.

This fix relied on request.getRequestURI() to end with a slash when
the startup URL had the form `/context/` (as opposed to `/context`).
Unfortunately, this does not work when the server uses URL rewriting
for session management. The resulting error is worse than the issue
the this fix tried to work around.

See bug 429443: Requests to Root Path Entry Point are not redirected
correctly in WAR deployment
https://bugs.eclipse.org/bugs/show_bug.cgi?id=429443

Change-Id: I4f14f307590e67ecd71dd28a3aea6c0166535626
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>