Remove unused support for custom ID generators
This internal support had been implemented in an effort to provide an
alternative to setting custom widget IDs using setData.

However, since the RemoteObject must be created in the constructor of
a widget already, a custom ID generator would have to be called with
an incomplete object, which is inherently unsafe.

A custom widget id generator that cannot access any methods on the
widget to generate an ID for cannot do much more than the standard
id generator.

See bug 392526: Allow a pluggable custom id generator (for testing)
https://bugs.eclipse.org/bugs/show_bug.cgi?id=392526

Change-Id: Ic61977ed1b8d53d3ed29b13663877b58b2615d34
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>