Remove localrepo from examples build
The p2 repository 'localrepo' had been used to simplify the process of
adding optional/additional bundles to the examples demo, but has never
been popular. Removing this localrepo now because it introduces build
failures when using spaces in paths on certain platforms.

Bug 425280: WAR build failed if there are spaces in repo path
https://bugs.eclipse.org/bugs/show_bug.cgi?id=425280

Change-Id: Iafeedf5c211a5fb39b281686f0a8057689989bcc
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>