Rework client-side Menu focus behavior
To fix "Bug 426570 - IFocusService and its focus tracking is not working
anymore", the menu and menubar focus behavior as been drastically
changed. Instead of always returning the focus to it's previous owner,
menus and menu bars are now impossible to focus. As a result there will
never be a focus event fired by using a menu, and the currently focused
widget will keep the focus theming all the time.

To ensure that the menu is still navigateable by keyboard, a menu now
sets "capture" to true to recieve all key events while it is on top.
This also fixes a minor issue where menus did not react to keyboard
input in some situations.

While this makes the code smaller, there is the small downside that
mouse event's are also captured and have to be re-dispatched if they do
not belong to the menu.

This commit also does a number of smaller refactorings on the menu,
including deleting some dead code. The tests have also been adjusted,
extended and in some cases split to make them shorter. The menu bar now
has it's own test class.

Change-Id: Ica5cf1a9d1385e72b68efb666a924115d775753a
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>