Fix randomly failing browser javascript tests
Browser tests could fail for a number reasons.

- Browser needed to long to load a URL : Not entirely fixed, but
increased the delay in ciritical places. Long term solution should be to
wait for the load event instead of a fixed time span.

- In webkit the test for access to an restriced iframe could fail
because Webkit recently introduced a bug or feature where instead of
causing an exception the wrong document would be accessed. This is now
fixed by checking the domain of the document.

- For some reason IE sometimes throws an error when attempting to delete
a global variable, as happens in BrowserTest. It's now assigned
undefined instead. The value of this test is questionable, we should
consider to delete it.

- In older IE the tests checking that no error occured while waiting for
the iframe to load could fail due to still unloading iframes from
previous tests causing errors when attempting to send progress events.
These errors are now prevented by checking if the browser has an
ObjectRegistry entry before sending the event.

This last change required some updates in the fixture - properly
clearing the object registry and removing listener for sending requests.

Change-Id: I1c3dd8c0a0d4b907cdc99117bef7270e31e8f61a
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>