Prevent absolute path in startup message url property
Tomcat's implementation of HttpServletResponse.encodeURL() turns the
empty string (which is the servlet name in case of using the root path)
into an absolute URL. This interferes with URL rewriting.

The solution is to use "./" instead of the empty string.

See bug 430285: Root path don't not work behind URL rewriting proxy
https://bugs.eclipse.org/bugs/show_bug.cgi?id=430285

Change-Id: Ifdae4d782ca82b123df2c761c208b64a0e904a64
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>