Change JavaScript Popup widget focus behavior
Most of the widgets (Menu, ToolTip, WidgetToolTip, DropDown popup) that
implement rwt.widgets.base.Popup are not focusable - only Shell (Window)
is.
With this change now popups are either a focus roots themselves, or 
don't have a focus root at all (making them and their children not 
focusable at all). It's no longer possible for a popup to be a focused
child of the ClientDocument.
Make Popup#_isFocusRoot false by default and override it in the
Window.js only.

Change-Id: Ie378491746d9ec229bb595901503356482218d4c
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>