Prevent crash when preparing enhanced (SVG/VML) background
See Bug 432909 - Client may crash in IE7/8/9 due to implementation error
in Parent._computeVisibleChildren

Possibly introduced by fixing Bug 359665 - Background transparent don't
work in IE 

It is possible under some unknown (but rare) conditions to crash the
client if a widget is creating an SVG/VML background (calling
prepareEnhancedBorder) while it's children are not yet visible but
already returned as such by _computeVisibleChildren.

The implementation of _computeVisibleChildren, while technically not
correct, can not be changed because some methods in Parent.js rely on
it's specific behavior. Therefore I added a pre-check just before the
expression that caused the crash.

Change-Id: I432ee325f3925635395be1a6f4170ddea4479527
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>