Remove ProtocolUtil.getXxxPropertyValueAsString() methods
These methods convert JsonValues to Java values and call toString() on
the result. This type-guessing is needless as internal code knows the expected types.

Rework callers of this method to use JsonMapping explicitly.

Move implementation and tests to the deprecated public WidgetLCAUtil
to keep its API intact and remove the corresponding methods from the internal WidgetLCAUtil.

Change-Id: I7a4f8a0525ac39b3d41e63658755f2f4fd329c10
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>