Cleanup JavaScript Colors util
Moving utility methods from the Colors object into a surrounding
function scope helps to save almost 400 bytes in the compressed JS.
Probably this is because the compressor treats the names of
underscore-prefixed methods as API and does not minify them.

Remove CSS2 system colors (such as "buttonface"). These names are not
used in our code and marked as deprecated by the CSS spec anyway. Also
remove non-standard color keywords.

Change-Id: I7fc04d27364ac732f819469e48cd5d61b1a0d695
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>