Eliminate qx bootstrapper
rwt.runtime.Bootstrap was used to create an rwt.qx.Class dummy in order
to define qx classes before the real qx.Class is loaded. Now that all
dependecies of qx.Class are plain objects, the bootstrapper is not
needed anymore.

Turn rwt.qx.Class into a plain object.

Create a file rwt.js that defines the object rwt and the namespace
function.

Move rwt utils to the top.

Change-Id: I943307d332c97abf86e13a11cfc49bc0049f59b6
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>