Always render Folder listener
CTabFolder has a single Folder listener for multiple events (min, max,
close, showList...). With change 22245, we suppressed notify operations
on the client (on remote object level) if there is no listener attach.
In case of "close" and "showList" events we must always send notify
operations (for proper layouting and so from the server-side). That's
why we always need Folder listener to be rendered.

433182: [CTabFolder] Click on >> does not work
https://bugs.eclipse.org/bugs/show_bug.cgi?id=433182

Change-Id: Ia78bea30c49e9b2045bd5251be49e509f475c015
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>