Translate "undefined" textColor to "inherit" in List.appearance.js
When color property is defined in CSS as "inherit", PropertyResolver
translate it as "transparent" and render it as "undefined" string. (see
ThemePropertyAdapterRegistry#ColorPropertyAdapter).
On the client, the "undefined" color property isn't recognized and need
translation again - for backgroundColor to "transparent", for textColor
to "inherit".

Remove unneeded background-color CSS List-Item definitions from default
theme.

394083: [List] Programmatically set foreground is lost on mouse out
https://bugs.eclipse.org/bugs/show_bug.cgi?id=394083

Change-Id: I77d7cb73dbcc491301216e4d62d06e13a17e2f4a
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>