Shutdown UISession immediately in case of an exception
In LifeCycleServiceHandler, we used to mark a UISession as started even
in case of an error during request processing. This marker ensured that
a session restart was detected in the next request and the current
UISession was properly shutdown.

However, when an unhandled exception occurs during the request
processing, the client receives an HTTP 500 and it is not possible to
continue this UISession. Hence, there is no reason to keep the UISession
alive and it's better to close it immediately.

Change-Id: Ie536e98641f0f2a2e10f9846b589713af2af8d24
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>