Pass ClientMessage instead of storing it
In an effort to change the message processing to be more functional-
style, pass the message to the functions that read from it in
LifeCycleServiceHandler. This allows for testingt these functions
directly, without faking messages in the request.

Only the RWTMessageHandler still puts the message in the service store
for the LCAs. Once we handle operations directly, the message does not
need to be kept in the service store anymore.

Change-Id: I287e6a457dd3ed89f740604a5ad53272b581d342
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>