Introduce MessageFilter
Message filters allow adding pre- and post-processing steps for
messages that are received from the client. The design follows the
same pattern known from servlet filters. All registered fiters build
up a chain. Every filter must delegate to the chain.

Internally, the chain elements are represented by instances of
MessageChainElement, containing a pointer to the filter and the
next element to delegate to.

MessageChainReference always points to the head of the chain, i.e.
the first chain element. RWTMessageHandler is the last element of
the filter chain, it actually processes the message.

Change-Id: I2da22eb235d1052f3d480fa55eda2b1bb75406d0
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>