Throw IllegalStateExcepion in Browser.execute if other script is pending
Since Browser.execute contains a main loop, it is possible for pending
SWT events to be dispatched before the method returns. (Usually if the
request contains multiple events.) If any of those events calls
Browser.evalaute it acted as if an JavaScript error occured
(ERROR_FAILED_EVALUATE), which is not correct. We are now firing an
IllegalStateException instead (in both evalute and execute), which makes
this scenario less puzzling and simpler to handle.

No test for that since there is no reasonable way to write any.

I also added the IllegalStateException to BrowserUtil.evalute to be
symmetrical.

See also Bug 371133 - [Browser] Support multiple script executions.

Change-Id: I50b536bdb58f922059010349b98c5354ec1f9953
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>