Distinguish RequestMessage and ResponseMessage
In order to understand the role of a Message object during request
processing, it seems to make sense to distinguish the types used for
request and response messages. This makes MessageFilters easier to
understand. With this distinction, it's no longer possible to return
the request message by mistake.

Also for some internal methods such as

    isRequestCounterValid( RequestMessage )

it makes sense to limit the method to request messages.

Both RequestMessage and ResponseMessage are subtypes of Message.
At the moment, both share the same API. However we may decide to
provide additional methods only for responses later.

Change-Id: I5e4f980e1c3354dea54b39dcd4a72b1fb4337efb
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>