Prevent client crash in DropDown#hasFocus
DropDown uses changeFocusedChild listener to hide itself when focus
control is changed. When executing DropDown#hasFocus (in a timer), it's
possible for getFocusRoot to return null (DropDown parent(s) is in
dispose and already disconnected from the widget hierarchy).

Change DropDown#hasFocus to return false if there is no focus root.

436757: Disposing TabFolder does not destruct TabFolder children on
Client side
https://bugs.eclipse.org/bugs/show_bug.cgi?id=436757

Change-Id: Ia2cc63cc08c9326bc2d654bbb39c5acf2fcf4ae5
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>