Increment versions of rwt.ui and related to 3.0
With the removal of the phaselistener extension point, the rap.ui
bundle's major version number must be incremented.

We've considered using the original version number for artifacts that
have been copied from the platform, however this needs more discussion.
For example, if we open up previously disabled API of the same
original version - we'd have to increment the minor version again...

Therefore increment the version to 3.0 for now, we can still bump to
3.7 later.

Change-Id: I75bbe6302cfd4f35d7f878cea5fe88ba2564c53d
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>