Introduce JS function rwt.define
Introduce rwt.define( name, object ) as a replacement for the global
namespace function and for rwt.qx.Class.createNamespace(). This
function can be used to create a namespace and to assign a given
object to a nested name. Example:

    rwt.define( "rwt.util.Strings", {

      contains: function( str, substr ) {
        ...
      },
      ...

    }

This avoids putting a function into the global namespace and it
eliminates the redundancy between namespace and
rwt.qx.Class.createNamespace().

Change-Id: I8176bb2270b954f2329f1f47628d5705b27a5248
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>