Add an option to expand the flexible cell in MultiCellWidget
GridColumnLabel (which extends MultiCellWidget) has a hack (override
_limitCellWidth function) to make flexible cell to grab the available
horizontal space and to push all cells on the right (sort icon in
column). This functionality will be needed also if MultiCellWidget is
used as a replacement of Atom by ExpandItem or CTabItem (push chevron
image in ExpandItem and close button in CTabItem to the right).

Add a combined getter and setter MultiCellWidget#expandFlexCell.
Expanded flexible cell completely occupies the available horizontal
space.

Change-Id: Ic47e20d0bd423a6950523e97db9ae9748c55ca97
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>