Use R-4.3.2-201402211700 as Eclipse Platform p2 repository
Because of a clean-up of the Eclipse Platform team we need to adjust the
location of the input p2 repository that we use as a target. The old
location at 4.3-M-builds/M-4.3.2RC3-201402061000/ doesn't exist any more
an is now replaced by the new released URL. The content that is used by
RAP is the same.

Change-Id: I8da5618aa7b590fad6a5e31339025ae6946b3c53
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>