Prevent jasmine from loading suites that are excluded by the filter
Jasmine lets you execute only specific suites/specs (for example by
clicking on the name in the reporter), but the file is still loaded,
parsed and reported (as skipped), which makes not that much faster than
just running all suites. As we now have a significant amount of suites,
especially when auto-converting the old tests to Jasmine suites
(?legacy=true), running only a few specs takes a unacceptably long time.
With these changes only Tests/Suites that are *potentially* executed
will be loaded by jasmine.html.

Change-Id: Ide5ec79abd55b98c2fba2af34b4b38a830531c99
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>