Remove remaining files from old javascript testrunner
All relevant code pieces have been incorporated into JasmineStartup.js.
Also some minor changes to the test runner: 

* The "TARGETENGINE" attribute for old-style tests is no longer
supported (as it is not required)
* The "filter" URL parameter is supported correctly again 
* The tests won't start if there are any errors logged during loading,
as that might cause a completely green suite even though not all tests
did run.
* The HtmlReporter can slow tests down considerably, especially in IE or
when the browser is in the background. With the URL paramter
"fast=true", the HtmlReporter is not created and only the browser
console is used.

Change-Id: Iee63047fec95eff66ca9edbfa80c1d2e2ec11588
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>