Declare UISession#exec as discouraged
Access to session singletons was the last use case for fake contexts.
Since SingletonUtil now provides methods that accept UISession or
ApplicationContext parameters, those singletons can read the
UISession / Application as parameters to their getInstance method.

There seems to be no use case that really requires UISession.exec().
I think it's too early to deprecate the method, but we should point
it out in the documentation.

Change-Id: I5783ea394d3a0cc251ac7b66f342aefe59fd18c7
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>