Refactor client-side ScrollBar/Slider code
These changes will be needed by later refactorings in Grid (which in
turn are necessary to implement new nebula grid features). First, the
scrollbar had an (internal) "minThumbSize" property that was moved to
AbstractSlider and the feature completeley re-implemented. The old
implemenation was very error-prone. Secont, the ScrollBar now has a
property "autoThumbSize", which is true by default (which causes the
same behavior as before). If set to false, setThumb can be used, which
will allow the Grid to scroll based on items instead of scrolling based
on pixels.

Change-Id: Ifbbb2f2a3b580992b254bf4bcc30c5d7b33cff66
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>