Add client-side support for TabItem badge
A "good" position for the badge is calculated based on the items padding
and badge size. The Item needs a margin that is at least half of the
fontsize of the badge, or the badge will be cut off by the TabFolder.
The badge is positioned slightly different on selected tabs because the
border width canges and there is .

Changes outside of Badges.js and TabItem are mostly related to
enhancents and improvements to RWTQuery.js.

Change-Id: Ie8318163c35aff2790cb6e05e18d6ce881542385
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>