Fix grid scrolling with touchpads/stepless scroll wheels
Traditional scroll wheels work like two buttons, one for 1 line up, one
for 1 line down. The Grid gets a +1 or -1 delta and scrolls accordingly
(2 rows). With touchpads or new scroll wheels designed for scrolling on
a pixel-basis the delta may be something like 0.2 (lines). This worked
fine for the previous scroll implementation which was based on pixels,
but now that it is based on items (rows/lines), it causes invalid states
in the scrollbar that may crash the server. With the change in this
commit the delta is alway rounded so at least one full line is scrolled.

Change-Id: I5d1b451e7ede425c80b68b94e5eff84c43f58a1d
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>