Adding an Eclipse Mars specific repository for Mars contribution
The Eclipse Mars Release will be using Eclipse Platform 4.5.0. With this
Mars specific p2 repository build it is possible to provide a clean RAP
Runtime p2 repository without version duplicates to this repository. It
will be used for Mars contributions only, RAP 3.0 will be based on and
published with Luna. 

Change-Id: Ic18c2cb2f567cb522329712a649003a57217b44d
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>