Fix jshint warnings and add Gruntfile for jshint-checks
Fix or add local jshint exceptions for all remaining jshint warnings in
org.eclipse.rap.rwt/js. Add a Gruntfile the the repository root to run
jshint against all these js files. The goal is to run grunt on the build
server and eventually use it to run the rwt javascript tests.

Change-Id: Iad9d340fac003ba3b2ba7de558841b5b13a11a87
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>