Fix Browser widget disabled state in IE9/10
When Browser is disabled, a "blocking" div is placed on top of the
iframe. This div should capture all mouse events and should not allow
them to propagate to the iframe.
It's a known issue in IE (9/10) [1] that a background image is needed on
a div in order to capture/fire mouse events.

Set empty background image on Browser blocking node for IE9/10.

[1] http://stackoverflow.com/questions/2158921/
empty-div-hover-event-not-firing-in-ie

441842: [Browser] Disabled state does not work in IE10
https://bugs.eclipse.org/bugs/show_bug.cgi?id=441842

Change-Id: I7d26b8cfd91d4b52cee3fa90316205006bf9ea57
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>