Provide JUnit rule as first official test API
Provide a class TestContext that provides a RAP environment in unit
tests. Instead of using Fixture.setUp() and Fixture.tearDown(), add
this statement to a Test:

    @Rule
    public TestContext context = new TestContext();

The Phase is set to PROCESS_ACTION to enable listeners.

In the future, we could add methods to the TestContext, e.g. to
inject mocks for Connection or RemoteObjects...

Change-Id: Ia05ab7b019ab4051a219e146b78d4c61635ce2a9
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>