Render Combo/CCombo selectionIndex when items changed
Current implementation of Combo/CComboLCA#renderSelectionIndex, based on
text monitoring, does not cover the following scenario:
- combo.setItems( new String[] { "a", "b" } );  combo.select( 0 );
-- in a subsequent request --
- combo.setItems( new String[] { "a" } );  combo.select( 0 );
In this case, the selection index is not rendered to the client as both
selectionIndex and text are unchanged.

Check for items change instead of text for proper detection when
selection index needs to be rendered.

444128: ComboViewer#setSelection does not work always when input changes
https://bugs.eclipse.org/bugs/show_bug.cgi?id=444128

Change-Id: Ida3e214ca2c8291928ef26e5af28660ed4fad0a1
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>