Convert GridRow from a qooxdoo widget to a simple HTML element wrapper
This was done so the browser will do the placement of the rows. Since
they are just stacked on top of each other it's a very simple task. As a
result resizes of rows cost much less time, and even animating their
height or offsets with CSS3 is possible.

All other changes result form this, including adapting tooltip support
and drag and drop support.

Change-Id: Iec37c52e8ad512f81f61c1cb129d5a163a0bb625
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>