Don't set text/selection on disposed Text/Combo/CCombo
Text and selection of Text/Combo/CCombo (with verify listener attached)
are not applied directly when processing the incoming message in the
coresponding operation handler. Setting text and selection is wrapped in
ProcessActionRunner to ensure correct event handling. When parent Shell
is closed in the same request where text/selection is set, "widget is
disposed" exception is thrown. In this case shell is closed first (all
child widgets are disposed) and then text/selection are set on already
disposed widget.

Add dispose check in Text/Combo/CComboOperationHandlers before setting
text and selection.

449350: Press keys then escape key fast enough in dialog generates
"Widget is disposed" exceptions
https://bugs.eclipse.org/bugs/show_bug.cgi?id=449350

Change-Id: I1075e6f2740fd4001eec17c8b9d3ba02d0508a82
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>