Don't send "modify" notify operation on disposed Text/Combo
Client-side Text/Combo widgets schedule a delayed (500ms) request when
"modify" notify operation need to be sent. It's possible to dispose the
widget (from the server-side) in this interval. In this case
Text/Combo._onSend function fail with JS error (remote object is not
registered as Text/Combo is already disposed).

449350: Press keys then escape key fast enough in dialog generates
"Widget is disposed" exceptions
https://bugs.eclipse.org/bugs/show_bug.cgi?id=449350

Change-Id: I65c3b8071db9882cf3a7ab9289e19dbc1b204e48
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>