Fix invalid value of Grid selection event index property
If Grid selection event index property is unknown, it should not be
included in the notify operation. In Grid.js#_fireSelectionChanged there
is a check if index is number, but this check does not work for null
values - isNaN( null ) returns false.

Use typeof instead of isNaN to make Grid.js#_fireSelectionChanged work
with null index value.

449432: Nebula grid: UnsupportedOperationException after click on RWT
hyperlink
https://bugs.eclipse.org/bugs/show_bug.cgi?id=449432

Change-Id: I393a71e15b55091241c4d4027e921795e7e464d7
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>