Animate scrollbar opacity change
Make scrollbars appear/disappear with a short fade effect. For the first
time we will use CSS3 transitions for this. The effect can not (yet) be
themed and may not work in all supported browser versions. 

Decrease appear delay to make up for the time the animation
needs. The appear delay could later be handled completele by the
transition effect, since it also supports a delay.

Change-Id: I62a574ff152bed53133bf3c26bd8235acb4a7089
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>