Fix class loader mocking in RWTServletContextListener test
To verify that contextInitialized() uses the context class loader, the
test mocks a class loader. When the ThemeManager reads resources during
activation, this class loader is requested to load some xml parser
classes. This leads to an error if the mocked class loader does not
delegate to the RWT class loader.

This error occurs only when running the test stand-alone, otherwise
the ThemeManager is already initialized by other tests.

Change-Id: I2e8379046d0a83faa9ed0026c04cd1f1d206f27d
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>