Fix client crash in case of missing row template cell position
The client crashes if both top/bottom or left/right of a cell are not
specified. This regression has been introduced by converting the cell
position properties in the protocol from number to array.

Check for existance of right/bottom properties before using them for
cell left and top calculation.

458142: [Template] Client crashes if cell position is not specified
correctly
https://bugs.eclipse.org/bugs/show_bug.cgi?id=458142

Conflicts:
	tests/org.eclipse.rap.rwt.jstest/js/org/eclipse/rwt/test/tests/TemplateTest.js

Change-Id: Ibb77694cd6c15d94ef6313c404d59237cce15707
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>