Render "data" property in the same operation with other properties
When control is created, all properties are rendered in "create"
operation. As "data" property is rendered after listeners, it's moved to
a separate "set" operation in case of available listeners. "data"
property can be used for setting the widget test id. Thus, it's
important to be in "create" operation (together with other control
properties).

Change-Id: Id308361612de56b0aa8ef953e60640c3546dc2ee