Eliminate the need of UISession in StartupJson
UISession is only needed in StartupJson to set the client error messages
based on currently used locale. As we are using client service
(ClientMessages) to render the localized messages, first GET request is
not the correct place to do it.

Add default English texts to all error messages on the client. These
messages will be shown only in case of initial startup request failure.
Move the rendering of client messages to UISessionBuilder and make
StartupJson UISession independent.

Change-Id: I9b67d347d9117d53039551d881610a33e874a8b1