Generate connection id on the server
To avoid CSRF vulnerability it's recomended to use synchronizer token.
We could use connection id (cid) as synchronizer token, but we need to
generate it on the server, not on the client.

Respect connection id in test fixture and cluster tests. It's not
possible to use null connection id anymore as it is generated during UI
session creation.

413668: Cross Site Request Forgery vulnerability (aka CSRF/XSRF)
https://bugs.eclipse.org/bugs/show_bug.cgi?id=413668

Change-Id: Iba88f49a3fed3a90b479fe92ebd98c421cb00e83