Fix wrong ScrollBar page increment calculation
For Tree and Table the scrolling is based on items, on ScrollComposite
on pixels. The page increment value was always calculated in pixels
(ScrollBar#_updatePageIncrement). As a result, the page incement was too
big (in pixels) for item based scrolling (Tree/Table).

Use thumb length as page increment value as it is dynamically adjusted
for different scrolling modes.

445607: [Table][Tree] Click on empty scrollbar space scrolls to the end
instead with page increment
https://bugs.eclipse.org/bugs/show_bug.cgi?id=445607

Change-Id: I2da900c6ba737d94cbc753af5c5e28eaa5746865