Don't use prefix when preserving listeners in WidgetLCAUtil
The "listener_" prefix was used to avoid collisions when preserving
property and listener with the same name. Since RAP 2.1 all properties
start with lower-case and all listeners with upper-case. The only
property/listener pair currently with similar name is
"selection/Selection".
Removing the prefix leads to better performance as we don't create new
strings all the time.

Change-Id: I5c7c678778ca0221a995f46646700c9c23ab202d