Fix wrong listener is removed if rap.off is called multiple times
rap.off is using rwt.util.Arrays.removeAt(array, index) to remove the
listener handler from the array. If handler is not registered index
parameter is -1. As this function internally uses "splice", negative
index leads to removal of an item from the end of the array, which is
wrong listener.

465413: Callback deregistration (rap.off) deletes wrong callback if
called multiple times
https://bugs.eclipse.org/bugs/show_bug.cgi?id=465413

Change-Id: Idb67ba88026c1950eb12763217051bb1ccee31db