Customized stack presentation for placeholder folder not working Bug 465646
Setting the ID of the real container enables the interaction design
implementation to find the corresponding stack presentation.

Basically, the problem originates in PartStack.presentationSite#getProperty(). 
The function redirects to PartStack#getID() if the property is "id". 
This "id" is used to find the corresponding stack presentation. 
If the id is not set for the real container of the container placeholder,
the stack presentation cannot be found.

This fix should not cause any problems. 
The ContainerPlaceholder is delegating to the real container. 
References of ContainerPlaceholder also show that the real container is used.

A similar pattern can be found in ViewSashContainer#restoreState().

Change-Id: Id74498b70c19b3495a91db5430d2b2cf62ce7da7
Signed-off-by: Georg Breitschopf <georg.breitschopf@gmail.com>