Prevent PDF to cover client-side error boxes
When PDF is opened in a Browser widget it covers everything including
client-side error boxes in case of network/timeout problems. This issue
mainly affects IE. To workaround it, load empty page (unload PDFs) in
all iframes available in the document before showing the error box. Do
this for all error types except connection error, where session is not
restarted, but last POST request is resent. 

Change-Id: I51fd744ffb1dafe30f6809eb9d11aa1c416bdd7f