Introduce type BoxDimensions for theming borders and padding
Using Rectangle for padding and border widths was problematic because

a) it does not contain fields for the right and bottom edge, these
   had been represented by (width - x) and (height - y), respectively.
b) it is a mutable type and cannot safely be shared.

Introduce a new type named BoxDimensions, that is immutable and
contains fields for the four edges. Use this type in the public
ControlThemeAdapter.

Change-Id: I02b06e326d849304e0c977040d75aa5c75969812