Introduce type Size for image sizes in theming
Using Point for image sizes was problematic because

a) its fields are named x and y instead of width and height, which made
   the consuming code harder to understand.
b) it is a mutable type and cannot safely be shared.

Introduce a new type named Size with fields width and height. Use this
type in ThemeAdapter methods that return sizes of theming images.

Size is immutable and its instances are kept in CssImage and can be
shared. Except a couple of places that should be reworked separately,
no new Size objects are created when obtaining image sizes from the
theme.

Change-Id: I213ef030a503c4e43cb7bbb8f513d8b39fcf8d40