Use BoxDimensions instead Rectangle in theme adapters
This is a follow-up re-factoring after commit
ab3133277ed57c6317d5be04b48832fa487d9e5a.
The BoxDimensions type is immutable and contains fields for the four
edges. BoxDimension objects are re-used and don't need to be created
again and again. Thus, the number of transient objects created during
request processing is reduced. 
The new type is now used as a return type for border, padding and margin
in all theme adapters.

Change-Id: I5fbaf1a8fc27c598b37c32ab3449c98d80eeef6d