Remove checks for processed message in test fixture
ProtocolUtil set a flag in the service store when the message from
the client was first accessed. This flag was used to protect Fixture
methods that modify the message from being called at the wrong time.

However, setting this flag in the session store affects runtime
performance even though it is only needed for tests. Since Fixture
is not API, we should skip this check. We should't need to defend
ourselves against badly written tests.

Change-Id: Iff4f9dfb1cf7590878f509e4e459aa868ba97b4e