Provide getter for ApplicationContext in ApplicationRunner
The ApplicationRunner creates an ApplicationContext and allows to
start and stop it. Since this class is basically wrapping the
ApplicationContext it seems fair to provide a getter.

As long as we don't provide API to obtain all servlet paths on
ApplicationContext, the rwt.osgi needs to access internal API,
but we get rid of the deprecated method in the rwt bundle.

Change-Id: I0857b234ff386e707e63778946ea3c83f84e8480