Eliminate theme warnings
We introduced warnings for incomplete themes as of bug 254478 [1].
Back then, properties missing in a custom theme were taken from the
default theme.

With the introduction of fallback themes [2], custom themes don't have
to define all properties anymore. Missing values are taken from the
fallback theme. Hence the check is not needed anymore.

As a side effect, the parameter `theme` in the CssValuesMap constructor
becomes obsolete and is removed.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=254478
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=363736

Change-Id: I6dfd054000bb437b4b39e3e2d8ea7792a1f12ed5