Rename WidgetAdapter to RemoteAdapter
This type of adapter is used for preserving property values and other
tasks related to the communication with the remote objects.

Change its name to reflect this purpose and allow for a hierarchy of
RemoteAdapters that does not clash with existing control adapters such
as ControlAdapter, CompositeAdapter, etc.

Rename the implementation to WidgetRemoteAdapter.

Add missing tests for WidgetRemoteAdapter.

Change-Id: I283457fd4a54b0473041bcb8797739d34616b6c3