Refactor ToolItem dropDown arrow implementation
The new dropDown arrow implementation does not use MultiCellWidget
cells, but HTML elements directly. This is done to prepare the ToolItem
for vertical orientation support. 

MultiCellWidget.js is modified to give ToolItem a hook to adjust the
space available for layouting cells (subtracting the arrow area).

RWTQuery.js is modified to support the "offset" method now used in
ToolItem. 

TestUtil.js is modified to let getElementBounds() support percentages
(provided the element is connected to the DOM.

Update theming to use a transparent ToolItem border instead of a "none",
otherwise the arrow would shift when the item is hovered.

Change-Id: I638ca9b7fdd2e2924834371823ca22a67bfd91de
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>