Add vertical orientation support to MultiCellWidget
This will be used to support ToolItems with icons above the text instead
of side by side.

The methods in MultiCellWidget that previosuly handled the horizontal
layout will now handle the layout for whatever is the content "flow"
direction. (Vertical if "vertical" property is true, otherwise
horizontal.) The methods that handled the vertical layout will now
handle the direction orthogonal to the flow direction.

The methods dedicated to computing the total spacing between cells have
been removed since this value is now computed in _getContentSizeFlowDir.

Change-Id: I8018d552648c66c75ff6c30900440f3ad6c959c6
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>