Remove bundles org.eclipse.rap.junit and org.eclipse.rap.junit.runtime
These projects have been created to execute PDE tests with RAP. We
haven't seen much interest in this approach and don't know of any
consumers.

As these projects are not actively maintained anymore and depend on
JUnit3, we are not including them in RAP 3.0. In addition to the project
removal, this change updates the build (pom, SDK feature, launch
configs) to match the new configuration.

The last working version is part of the 2.3-maintenance branch.

Bug 467181 - Remove bundles org.eclipse.rap.junit and
org.eclipse.rap.junit.runtime
https://bugs.eclipse.org/bugs/show_bug.cgi?id=467181

Change-Id: Id5de22b4d46b11a65ad8a1f2680ec32c9360b2aa
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>